refactor(ae2): Replace module with namespace, extend component map
