+++ examples/ae2.ts
@@ -1,0 +1,4 @@
+import * as component from "component";
+
+const { me_controller } = component;
+me_controller.store();
