+++ packages/AE2/component_map.d.ts
@@ -1,0 +1,8 @@
+declare namespace OC {
+    interface ComponentMap {
+        me_controller: ME.Controller;
+        me_interface: ME.Interface;
+        me_importbus: ME.ImportBus;
+        me_exportbus: ME.ExportBus;
+    }
+}
