+++ packages/openos/package.json
@@ -14,7 +14,6 @@
     "license": "MIT",
     "author": "Exeteres",
     "dependencies": {
-        "lua-types": "^2.10.1",
         "@opct/common": "^0.0.1"
     }
 }
